code optimized for MainActivity.kt add getActivityLauncher, changes in initListener
