--- app/src/main/java/com/example/notesapp/MainActivity.kt
@@ -1,119 +1,0 @@
-package com.example.notesapp
-
-import android.app.Activity
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.view.MenuItem
-import android.widget.LinearLayout
-import android.widget.PopupMenu
-import android.widget.SearchView
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.cardview.widget.CardView
-import androidx.lifecycle.ViewModelProvider
-import androidx.recyclerview.widget.StaggeredGridLayoutManager
-import com.example.Adapter.NotesAdapter
-import com.example.ViewModel.NotesViewModel
-import com.example.data.Note
-import com.example.data.NoteDatabase
-import com.example.notesapp.databinding.ActivityMainBinding
-
-class MainActivity : AppCompatActivity(), NotesAdapter.NotesItemClickListener,
-    PopupMenu.OnMenuItemClickListener {
-    private lateinit var binding: ActivityMainBinding
-    private lateinit var database: NoteDatabase
-    private lateinit var viewModel: NotesViewModel
-    lateinit var notesAdapter: NotesAdapter
-    lateinit var selectedNote: Note
-
-    private val updateNote =
-        registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
-            if (result.resultCode == Activity.RESULT_OK) {
-                val note = result.data?.getSerializableExtra("note")
-                if (note != null) {
-                    viewModel.updateNote(note as Note?)
-                }
-            }
-        }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        initUI()
-        viewModel = ViewModelProvider(
-            this,
-            ViewModelProvider.AndroidViewModelFactory.getInstance(application)).get(NotesViewModel::class.java)
-        viewModel.allNotes.observe(this) { list ->
-            list?.let {
-                notesAdapter.updateList(list)
-            }
-        }
-
-        database = NoteDatabase.getDatabase(this)
-    }
-
-    private fun initUI() {
-        binding.recyclerView.apply {
-            setHasFixedSize(true)
-            layoutManager = StaggeredGridLayoutManager(2, LinearLayout.VERTICAL)
-            adapter = NotesAdapter(context, this@MainActivity)
-            adapter = notesAdapter
-        }
-        val getContext =
-            registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
-                if (result.resultCode == Activity.RESULT_OK) {
-                    val note = result.data?.getSerializableExtra("note") as? Note
-                    if (note != null) {
-                        viewModel.insertNote(note)
-                    }
-                }
-            }
-        binding.fbAdd.setOnClickListener {
-            val intent = Intent(this, AddNote::class.java)
-            getContext.launch(intent)
-        }
-
-        binding.searchView.setOnQueryTextListener(object : SearchView.OnQueryTextListener {
-            override fun onQueryTextChange(newText: String?): Boolean {
-
-                return false
-            }
-
-            override fun onQueryTextSubmit(newText: String?): Boolean {
-                if (newText != null) {
-                    notesAdapter.filterList(newText)
-                }
-                return true
-            }
-        })
-    }
-
-    override fun onItemClicked(note: Note) {
-        val intent = Intent(this@MainActivity, AddNote::class.java)
-        intent.putExtra("current_note", note)
-        updateNote.launch(intent)
-    }
-
-    override fun onLongItemClicked(note: Note, cardView: CardView) {
-        selectedNote = note
-        popUpDisplay(cardView)
-    }
-
-    private fun popUpDisplay(cardView: CardView) {
-        val popup = PopupMenu(this, cardView)
-        popup.setOnMenuItemClickListener(this@MainActivity)
-        popup.inflate(R.menu.pop_menu)
-        popup.show()
-    }
-
-    override fun onMenuItemClick(item: MenuItem?): Boolean {
-
-        if (item?.itemId == R.id.delete_note) {
-            viewModel.deleteNote(selectedNote)
-            return true
-        }
-        return false
-    }
-}
